Trying to debug

+++ Assets/Scripts/Inventory/Items.cs
@@ -6,10 +6,10 @@ using UnityEngine;
     [CreateAssetMenu(fileName = "Item", menuName = "Item/Create New Item")]
     public class Items : ScriptableObject
     {
-        [Header("Item Information")]
-        public Sprite itemIcon;
-        public string itemName;
-        public int itemId;
+    [Header("Item Information")]
+    public Sprite itemIcon;
+    public string itemName;
+    public int itemId;
 
     public bool isArmor = false;
     public bool isWeapon = false;
@@ -19,21 +19,21 @@ using UnityEngine;
     [Header("ONLY CHANGE IF ITEM IS ARMOR")]
     //additive
     [Header("Defense")]
-    public float defense = 0;
+    public int defense = 0;
 
     //additive
     [Header("Health")]
-    public float health = 0;
+    public int health = 0;
 
     //multiplicative
     //it times by the value set by the class
     [Header("Movement Speed")]
-    public float movespeed = 0;
+    public int movespeed = 0;
 
     //multiplicative
     //it times by the value set by the class
     [Header("Jump Height")]
-    public float jumpforce = 0;
+    public int jumpforce = 0;
 
     //additive
     [Header("Jump Modifier")]
